Add SliderAutomationPeer
