+++ src/Windows/Avalonia.Win32/Automation/RootAutomationNode.cs
@@ -12,8 +12,11 @@ namespace Avalonia.Win32.Automation
 {
     [RequiresUnreferencedCode("Requires .NET COM interop")]
     internal class RootAutomationNode : AutomationNode,
-        IRawElementProviderFragmentRoot
+        IRawElementProviderFragmentRoot,
+        IRawElementProviderAdviseEvents
     {
+        private int _raiseFocusChanged;
+
         public RootAutomationNode(AutomationPeer peer)
             : base(peer)
         {
@@ -42,6 +45,36 @@ namespace Avalonia.Win32.Automation
         {
             var focus = InvokeSync(() => Peer.GetFocus());
             return GetOrCreate(focus);
+        }
+
+        void IRawElementProviderAdviseEvents.AdviseEventAdded(int eventId, int[] properties)
+        {
+            switch ((UiaEventId)eventId)
+            {
+                case UiaEventId.AutomationFocusChanged:
+                    ++_raiseFocusChanged;
+                    break;
+            }
+        }
+
+        void IRawElementProviderAdviseEvents.AdviseEventRemoved(int eventId, int[] properties)
+        {
+            switch ((UiaEventId)eventId)
+            {
+                case UiaEventId.AutomationFocusChanged:
+                    --_raiseFocusChanged;
+                    break;
+            }
+        }
+
+        protected void RaiseFocusChanged(AutomationNode? focused)
+        {
+            if (_raiseFocusChanged > 0)
+            {
+                UiaCoreProviderApi.UiaRaiseAutomationEvent(
+                    focused,
+                    (int)UiaEventId.AutomationFocusChanged);
+            }
         }
 
         public void FocusChanged(object? sender, EventArgs e)
